+++ src/ViewHoldingsRecord.js
@@ -132,9 +132,11 @@ class ViewHoldingsRecord extends React.Component {
 
   updateHoldingsRecord = (holdingsRecord) => {
     const holdings = holdingsRecord;
+
     if (holdings.permanentLocationId === '') delete holdings.permanentLocationId;
     if (holdings.temporaryLocationId === '') delete holdings.temporaryLocationId;
-    this.props.mutator.holdingsRecords.PUT(holdings).then(() => {
+
+    return this.props.mutator.holdingsRecords.PUT(holdings).then(() => {
       this.onClickCloseEditHoldingsRecord();
     });
   }