+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add info callout when saving file with the instance UUIDs takes longer. Refs UIIN-1329.
 * Do not escape quotes when submitting query-search queries. Refs UIIN-1319, UIIN-1320.
 * Use `TextArea` for some data fields on instance form. Refs UIIN-1278.
+* Return promise when submitting instance, holdings record or item form. Fixes UIIN-1339 and UIIN-1340.
 
 ## [5.0.1](https://github.com/folio-org/ui-inventory/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.0...v5.0.1)